Added Menu System

+++ config/plugins.php
@@ -104,13 +104,7 @@ return [
                         'subPath'       => '/website/components',
                         'Name'          => 'Component',
                         'permission'    => 'Hardel.Website.Component'
-                    ],
-                    /*[
-                        'subPath'       => '/website/elements',
-                        'Name'          => 'Element',
-                        'permission'    => 'Hardel.Website.Element'
-                    ],*/
-
+                    ]
                 ],
                 'migration-up'      => function(){
                     //here migration to create tables and columns
@@ -131,18 +125,6 @@ return [
                         });
                     }
 
-                    /*if(! $Schema::hasTable('lt_elements'))
-                    {
-                        $Schema::create('lt_elements',function(Illuminate\Database\Schema\Blueprint $table){
-                            $table->increments('id');
-                            $table->string('name');
-                            $table->string('Object')->nullable();
-                            $table->string('function')->nullable();
-                            $table->text('appearance')->nullable();
-                            $table->timestamps();
-                        });
-                    }*/
-
                     if(! $Schema::hasTable('lt_components'))
                     {
                         $Schema::create('lt_components',function(Illuminate\Database\Schema\Blueprint $table){
@@ -153,37 +135,38 @@ return [
                         });
                     }
 
-                    if($Schema::hasTable('lt_components') and $Schema::hasTable('lt_pages'))
+                    if( $Schema::hasTable('lt_components') and $Schema::hasTable('lt_pages'))
                     {
-                        $Schema::create('lt_page_component',function(Illuminate\Database\Schema\Blueprint $table){
-                           $table->increments('id');
-                           $table->integer('idPage')->unsigned();
-                           $table->foreign('idPage')->references('id')->on('lt_pages')->onUpdate('cascade')->onDelete('cascade');
-                           $table->integer('idComponent')->unsigned();
-                           $table->foreign('idComponent')->references('id')->on('lt_components')->onUpdate('cascade')->onDelete('cascade');
-                           $table->string('Object')->nullable();
-                           $table->string('function')->nullable();
+                        if(! $Schema::hasTable('lt_page_component')) {
+                            $Schema::create('lt_page_component', function (Illuminate\Database\Schema\Blueprint $table) {
+                                $table->increments('id');
+                                $table->integer('idPage')->unsigned();
+                                $table->foreign('idPage')->references('id')->on('lt_pages')->onUpdate('cascade')->onDelete('cascade');
+                                $table->integer('idComponent')->unsigned();
+                                $table->foreign('idComponent')->references('id')->on('lt_components')->onUpdate('cascade')->onDelete('cascade');
+                                $table->string('Object')->nullable();
+                                $table->string('function')->nullable();
+                            });
+                        }
+                    }
+
+                    if( !$Schema::hasTable('lt_menus') and $Schema::hasTable('lt_pages')) {
+
+                        $Schema::create('lt_menus',function(Illuminate\Database\Schema\Blueprint $table){
+                            $table->increments('id');
+                            $table->string('name');
+                            $table->integer('idPage')->unsigned()->nullable();
+                            $table->foreign('idPage')->references('id')->on('lt_pages')->onUpdate('cascade')->onDelete('cascade');
+                            $table->integer('idParent')->unsigned();
                         });
                     }
 
-                    /*if($Schema::hasTable('lt_components') and $Schema::hasTable('lt_elements'))
-                    {
-                        $Schema::create('lt_component_element',function(Illuminate\Database\Schema\Blueprint $table){
-                            $table->increments('id');
-                            $table->integer('idComponent')->unsigned();
-                            $table->foreign('idComponent')->references('id')->on('lt_components')->onUpdate('cascade')->onDelete('cascade');
-                            $table->integer('idElement')->unsigned();
-                            $table->foreign('idElement')->references('id')->on('lt_elements')->onUpdate('cascade')->onDelete('cascade');
-                            $table->integer('idPadre')->unsigned();
-                        });
-                    }*/
                 },
                 'migration-down'    => function(){
                     //here migration to delete tables
                     $Schema=  Illuminate\Support\Facades\Schema::class;
-                    //$Schema::dropIfExists('lt_component_element');
                     $Schema::dropIfExists('lt_page_component');
-                    //$Schema::dropIfExists('lt_elements');
+                    $Schema::dropIfExists('lt_menus');
                     $Schema::dropIfExists('lt_components');
                     $Schema::dropIfExists('lt_pages');
                 },
